requirements.txt
django-registration-redux==2.6
pbr==5.4.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
virtualenv==16.6.1
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
social-auth-core
social-auth-app-django
python-social-auth==0.2.7
django-bootstrap3==5.4.0
django-bootstrap-themes==3.1.2
django-debug-toolbar